Share CPU line-sweep bake code between line-sweep and denali notebooks
Both notebooks had their own near-identical copy of the CPU sweep core
and bake helpers, so the tile-boundary fixes had to be applied twice and
would drift. Hoist the shared CPU code into src/lib so a single source of
truth serves both (imported through each notebook's existing ./lib
symlink):

  src/lib/line-sweep-core.js  — sweepCore + sweepCoreSource. Denali's
      copy was already a strict superset (optional horizonPN, defaulting
      to W so line-sweep behaviour is unchanged), so it becomes the
      canonical file; git tracks it as a rename of denali's sweep-core.js.
  src/lib/line-sweep-bake.js  — the per-tile bake helpers (buildRowPx,
      computeNormals, computeLSAO, computeShadow, packRGBA, normInv,
      AO_DIRECTIONS/AO_GAMMA), which were byte-identical between the two
      workers apart from the optional horizonPN arg.

The two Web Worker files keep only their (genuinely different) message
protocols and now import the computation from the shared lib:
  - line-sweep cpu-bake-worker.js (type-tagged protocol)
  - denali bake.worker.js (WorkerPool id-based protocol)
The other line-sweep CPU importers (sweep-shadow, lsao-sweep, soft-sweep)
repoint to ./lib/line-sweep-core.js. Both per-notebook sweep-core.js
copies are deleted.

Only the CPU path is shared; each notebook's experimental GPU
webgpu-pipelines.js is left untouched (denali doesn't use it). Net −699
lines. Verified: both worker import graphs bundle through the symlinks,
the full vite transform resolves all 641 modules, and the seam harness
against the shared core reproduces the fixed exit-edge error (0.004-0.006
vs the old 0.0345 clamp).
